requirements.txt
pydantic>=2.0,<3.0
fastapi>=0.110.0
uvicorn>=0.27.0
anthropic>=0.40.0
agentscope>=0.1.0
pytest>=8.0.0
websockets>=12.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
mem0ai>=0.1.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0
